requirements/base.txt
django-environ==0.11.2
pandas == 2.0.3
matplotlib == 3.7.5